Numerous changes to the Java code, and example
Most notably:

- switch away from bytes and shorts in the public interface (because
they're a bit annoying to use - having to cast all the time)

- add a GreeterClient wrapper around the raw GRPC structures

- add a GreeterServer implementation of the Greeter service

- give the two above classes main methods, and get rid of the Example class

+++ examples/java/src/main/java/helloworld/LoadBalancer.java
@@ -23,10 +23,10 @@ class LoadBalancer implements Listener {
         this.random = new Random();
     }
 
-    public <T> Service<T> registerService(byte serviceByte, ServiceFactory<T> factory) {
-        assert !serviceFactories.containsKey(serviceByte);
-        serviceFactories.put(serviceByte, factory);
-        return new Service<>(serviceByte);
+    public <T> Service<T> registerService(int serviceByte, ServiceFactory<T> factory) {
+        assert !serviceFactories.containsKey((byte) serviceByte);
+        serviceFactories.put((byte) serviceByte, factory);
+        return new Service<>((byte) serviceByte);
     }
 
     public class Service<T> {
